Fixed display function for numbers above 15
The display function was not separating the 1st and 2nd digit properly
for displaying on the 7 segments. It was always displaying "22", "33",
"44", or "55". Now it can show you your score properly. Also added new
variable for total score kept in "total_time_microseconds". It will
display this more precise score on serial debug at 9600.