added display_counting_waiting() and changed LED reaction during gameplay
Now the waiting mode is showing the correct values and button combos
from 1-5. You can still hit a button to start a new game at any time.

Also, made it so the LEDs react to your pressing during the timeout
while loop. This is in an effort to show the player that they must hold
down the combo of buttons at the same time. It kind of shows that the
buttons don't change permanently because they only stay lit when you
press them.