Added randomSeed to setup, and COMBINED hex file
A customer, menehune23, commented on the product page about how the
first time you play after a power cycle the order of the sequence is
always the same. Adding a randomSeed in the setup (usuing A0 - before it
is setup with it's internal pullup) solved the issue. Now it is much
more random, and I doubt it will ever start with the same exact
sequence.

Also added the combined hex file that we are programming with in
production. This has been requested on other kits, so I thought I'd
throw it in this repo too.